ORM: move attributes/extras methods of frontend node to mixins
Move all methods related to attributes and extras from the frontend
`Node` class to separate mixin classes called `EntityAttributesMixin`
and `EntityExtrasMixin`. This makes it easier to add these methods to
other frontend entity classes and makes the code more maintainable.
